Modifications et ajout de la vérification des permissions sur les controlleurs

+++ EasyPDM/src/main/java/fr/upem/controller/UserManager.java
@@ -5,11 +5,16 @@
  */
 package fr.upem.controller;
 
+import fr.upem.easypdm.dao.implement.UseRoleDAO;
 import fr.upem.easypdm.dao.implement.UsersDAO;
 import fr.upem.easypdm.entity.Organisation;
 import fr.upem.easypdm.entity.Role;
+import fr.upem.easypdm.entity.UseRole;
 import fr.upem.easypdm.entity.Users;
-import fr.upem.security.RACWriter;
+import fr.upem.security.EntityType;
+import fr.upem.security.Operation;
+import fr.upem.security.RACs;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -19,41 +24,56 @@ import java.util.List;
 public class UserManager {
     
     private static final UsersDAO usersDAO = new UsersDAO();
-    private final Users user;
-    //private final ROAWriter roaWriter; 
+    private static final UseRoleDAO useRoleDAO = new UseRoleDAO();
+    private final RACs racs;
     
-    public UserManager(Users user) {
-        this.user = user;
-        //roaWriter = new ROAWriter()
+    public UserManager(AuthenticateBean session) {
+        racs = new RACs(session.getUser().getUseRoles());
     }
     
     public void addUser(Users user) {
-        
-        usersDAO.create(user);
+        if(racs.isPermitOperation(EntityType.USERS, Operation.CREATE, null)) {
+            usersDAO.create(user);   
+        }
     }
     
     public void removeUser(Users user) {
-        usersDAO.remove(user);
+        if(racs.isPermitOperation(EntityType.USERS, Operation.DELETE, null)) {
+            usersDAO.remove(user);   
+        }
     }
 
     public void updateUser(Users user) {
-        usersDAO.update(user);
+        if(racs.isPermitOperation(EntityType.USERS, Operation.MODIFY, null)) {
+            usersDAO.update(user);
+        }
     }
     
     public List<Users> getAllUsers() {
-        return usersDAO.findAll();
+        if(racs.isPermitOperation(EntityType.USERS, Operation.READ, null)) {
+            return usersDAO.findAll();   
+        }
+        return new ArrayList<>();
     }
     
     public List<Users> getOrganisationUsers(Organisation org) {
-        //TODO
-        return null;
+        if(racs.isPermitOperation(EntityType.USERS, Operation.READ, null)) {
+            //TODO DAO pour recuperer Users par Organisation
+        }
+        return new ArrayList<>();
     }
     
     public void addUserRoleInOrganisation(Users user, Role role, Organisation org) {
-        //TODO
+        if(racs.isPermitOperation(EntityType.USEROLE, Operation.CREATE, null)) {
+            useRoleDAO.create(new UseRole(user, org, role));
+        }
     }
     
     public Users getUser(long id) {
-        return usersDAO.find(id);
+        if(racs.isPermitOperation(EntityType.USERS, Operation.READ, null)) {
+            return usersDAO.find(id);   
+        }
+        // A voir pour la gestion des erreurs
+        return new Users();
     }
 }
